+++ ui/src/app/core/accessories/accessories.interfaces.ts
@@ -2,6 +2,7 @@ import type { ServiceType } from '@homebridge/hap-client'
 
 export type ServiceTypeX = ServiceType & {
   customName?: string
+  customType?: string
   hidden?: boolean
   onDashboard?: boolean
 }