+++ CHANGELOG.md
@@ -33,6 +33,10 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - add plugin support button and modal
 - add plugin search button, hide bar by default
 - add tab in plugins showing plugin analytics in `iframe`
+- allow applying a custom service to an accessory's service
+  - change between `Switch`, `Outlet`, `Lightbulb`, `Fan` for these accessory types
+  - change between `Door`, `Window`, `WindowCovering` for these accessory types
+  - this can be set in the accessory info modal, more options to come in the future
 
 ### Other Changes